+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -9,7 +9,6 @@ use App\Http\Requests\Api\V1\UpdateTicketRequest;
 use App\Http\Resources\V1\TicketResource;
 use App\Models\Ticket;
 use App\Policies\V1\TicketPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class TicketController extends ApiController
@@ -40,14 +39,12 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request)
     {
-        try {
-            //policy
-            $this->isAble('store', Ticket::class);
-
+        //policy
+        if ($this->isAble('store', Ticket::class)) {
             return new TicketResource(Ticket::create($request->mappedAttributes()));
-        } catch (AuthorizationException $exception) {
-            return $this->error('Current user is not authorized to create this resource', 403);
         }
+
+        return $this->error('Current user is not authorized to create this resource', 403);
     }
 
     /**
@@ -90,16 +87,15 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             //policy
-            $this->isAble('update', $ticket);
+            if ($this->isAble('update', $ticket)) {
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            return new TicketResource($ticket);
-
+            return $this->error('Current user is not authorized to update this resource', 403);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $exception) {
-            return $this->error('Current user is not authorized to update this resource', 403);
         }
     }
 
@@ -112,12 +108,13 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             //policy
-            $this->isAble('replace', $ticket);
+            if ($this->isAble('replace', $ticket)) {
+                $ticket->update($request->mappedAttributes());
 
-            $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            return new TicketResource($ticket);
-
+            return $this->error('Current user is not authorized to update this resource', 403);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
         }
@@ -132,11 +129,13 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             //policy
-            $this->isAble('delete', $ticket);
+            if ($this->isAble('delete', $ticket)) {
+                $ticket->delete();
 
-            $ticket->delete();
+                return $this->ok("Ticket $ticket_id deleted.");
+            }
 
-            return $this->ok("Ticket $ticket_id deleted.");
+            return $this->error('Current user is not authorized to delete this resource', 403);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
         }
